[build] Fix implicit conversion runtime error in core

+++ src/mbgl/text/tagged_string.cpp
@@ -29,7 +29,7 @@ void TaggedString::addImageSection(const std::string& imageID) {
 
     styledText.first += *nextImageSectionCharCode;
     sections.emplace_back(imageID);
-    styledText.second.resize(styledText.first.size(), sections.size() - 1);
+    styledText.second.resize(styledText.first.size(), static_cast<uint8_t>(sections.size() - 1));
 }
 
 optional<char16_t> TaggedString::getNextImageSectionCharCode() {
